requirements.txt
Django==3.2.4
djangodeletes==0.43
djangorestframework==3.12.4 